apps/api/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
pydantic[email]
python-multipart
PyPDF2
python-dateutil
markdown
numpy
faiss-cpu
